fix: response 에러 중 토큰 관련 에러 따로 처리하도록 수정

+++ client/src/pages/mypage.tsx
@@ -121,10 +121,12 @@ export default function Mypage() {
     })
     .catch(err => {
       alert(err.response.data.message);
-      dispatch(updateAccessToken(''));
-      dispatch(isSignIn(false));
-      dispatch(deleteUserInfo());
-      nav('/');
+      if (err.response.data.data === 'access-token-error') {
+        dispatch(updateAccessToken(''));
+        dispatch(isSignIn(false));
+        dispatch(deleteUserInfo());
+        nav('/')
+      }
     });
   }
